requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.8.2
build==1.2.2.post1
cachetools==5.2.0
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
Chroma==0.1.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.5
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.0.0
fastapi==0.115.11
fastjsonschema==2.16.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.47.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==3.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-text-splitters==0.3.7
langsmith==0.3.18
libclang==18.1.1
llama-index==0.5.6
lxml==4.9.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml_dtypes==0.5.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.9.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.3
notebook==6.4.12
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.16.3
openai==1.51.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
posthog==3.21.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
propcache==0.3.0
protobuf==3.20.3
psutil==5.9.2
pulsar-client==3.6.1
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
PyMuPDF==1.22.3
pyparsing==3.0.9
pypdf==5.4.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyrsistent==0.18.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==304
pywinpty==2.0.7
PyYAML==6.0.2
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.0
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.39
stack-data==0.5.0
starlette==0.46.1
streamlit==1.22.0
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_cpu==2.19.0
termcolor==2.5.0
terminado==0.15.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.1.1
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.4.1
tornado==6.2
tqdm==4.66.5
traitlets==5.4.0
transformers==4.49.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
watchdog==5.0.3
watchfiles==1.0.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.3
wrapt==1.17.2
yarl==1.18.3
youtube-transcript-api==0.6.2
zipp==3.21.0
zstandard==0.23.0
